add simple header
