+++ src/sddekit_simple.h
@@ -1,0 +1,16 @@
+
+
+/**  SDDEKit, simplified header.
+ */
+
+/* included for bool, uint32_t & FILE types */
+#include <stdbool.h>
+#include <stdint.h>
+#include <stdio.h>
+
+/* ignore these */
+#define SD_API
+#define restrict
+
+/* contains actual C declarations of API */
+#include "sddekit_api.h"
